restructure files for packaging with a setup.py
