+++ bin/virtme-init
