+++ bin/virtme-mkinitramfs
@@ -7,9 +7,9 @@
 # 8177f97513213526df2cf6184d8ff986c675afb514d4e68a404010521b880643
 
 import argparse
-import modfinder
-import virtmods
-import mkinitramfs
+import virtme.mkinitramfs
+import virtme.modfinder
+import virtme.virtmods
 
 _ARGPARSER = argparse.ArgumentParser(
     description='Generate initramfs image for virtme',
@@ -21,13 +21,13 @@ def main():
 
     args = _ARGPARSER.parse_args()
 
-    config = mkinitramfs.Config()
+    config = virtme.mkinitramfs.Config()
 
     if args.mod_kversion is not None:
-        config.modfiles = modfinder.find_modules_from_install(
-            virtmods.MODALIASES, kver=args.mod_kversion)
+        config.modfiles = virtme.modfinder.find_modules_from_install(
+            virtme.virtmods.MODALIASES, kver=args.mod_kversion)
 
-    mkinitramfs.mkinitramfs(sys.stdout.buffer, config)
+    virtme.mkinitramfs.mkinitramfs(sys.stdout.buffer, config)
 
     return 0
