+++ virtme/cpiowriter.py
